main/serializers/rentals.py
from rest_framework import serializers

from main.models import Rental


class RentalSerializer(serializers.ModelSerializer):
    class Meta:
        model = Rental
        fields = [
            "id",
            "name",
        ]
